Aleix38

+++ app/Http/Controllers/CU_43Controller.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Usuari;
+use App\Logs;
 use Krucas\Notification\Facades\Notification;
 use Illuminate\Support\Facades\DB;
 
@@ -27,10 +28,21 @@ class CU_43Controller extends Controller {
 
         $id = $request->cu43_idUsuari;
         $user = Usuari::findOrFail($id);
+        $nlog = Logs::where('idLog', $request->cu43_idLog)->first();
 
-        if ($user != null) {
+        if ($user != null && $nlog == null) {
+            
+            //Registrar Log
+            $nlog = new Logs;
+            $nlog->idUsuari =$id; 
+            $nlog->descripcio = "Eliminar Usuari";  
+            $nlog->dataLog = date('Y-m-d');
+            $nlog->hora = date('H:i:s');
+            $nlog->path = "";
+            $nlog->save();
+            
             $user->delete();
-
+            
             Notification::success("L'usuari s'ha eliminat correctament.");
             return redirect('CU_42_GestionarUsuaris');
         } else {
